+++ src/lookuptable.cpp
@@ -6,7 +6,7 @@ template <typename Value>
 LookupTable<Value>::LookupTable(const Value& default_value) : size_(0), default_value_(default_value) {}
 
 template <typename Value>
-void LookupTable<Value>::insert(int key, const Value& value) {
+void LookupTable<Value>::insert(unsigned short key, const Value& value) {
 	if (size_ < MaxSize) {
 		if (value != default_value_ && values_[key] == default_value_)
 			++size_;
@@ -15,7 +15,7 @@ void LookupTable<Value>::insert(int key, const Value& value) {
 }
 
 template <typename Value>
-Value& LookupTable<Value>::operator[](int key) {
+Value& LookupTable<Value>::operator[](unsigned short key) {
 	if (key >= 0 && key < MaxSize) {
 		return values_[key]; 
 	} else {  // Return default value if key is not found 
@@ -24,7 +24,7 @@ Value& LookupTable<Value>::operator[](int key) {
 }
 
 template <typename Value>
-const Value& LookupTable<Value>::operator[](int key) const { 
+const Value& LookupTable<Value>::operator[](unsigned short key) const { 
 	if (key >= 0 && key < MaxSize) { 
 		return values_[key]; 
 	} else { // Return default value if key is not found 
@@ -33,7 +33,7 @@ const Value& LookupTable<Value>::operator[](int key) const {
 }
 
 template <typename Value>
-bool LookupTable<Value>::has_key(int key) {
+bool LookupTable<Value>::has_key(unsigned short key) {
 	return table[key] != default_value_;
 }
